+++ .gitignore
@@ -1,2 +1,2 @@
-*.txt
 *.env
+*.txt
